WIP - package split
