+++ Makefile
@@ -53,11 +53,12 @@ uninstall:
 	$(call UNLINK_LAB_EXTENSION,@elyra/services)
 	$(call UNLINK_LAB_EXTENSION,@elyra/ui-components)
 	$(call UNLINK_LAB_EXTENSION,@elyra/metadata-common)
+	$(call UNLINK_LAB_EXTENSION,@elyra/script-editor)
 	$(call UNINSTALL_LAB_EXTENSION,@elyra/theme-extension)
 	$(call UNINSTALL_LAB_EXTENSION,@elyra/code-snippet-extension)
 	$(call UNINSTALL_LAB_EXTENSION,@elyra/metadata-extension)
 	$(call UNINSTALL_LAB_EXTENSION,@elyra/pipeline-editor-extension)
-	$(call UNINSTALL_LAB_EXTENSION,@elyra/script-editor-extension)
+	$(call UNINSTALL_LAB_EXTENSION,@elyra/python-editor-extension)
 	pip uninstall -y jupyterlab-git
 	pip uninstall -y jupyter-lsp
 	- jupyter labextension uninstall @krassowski/jupyterlab-lsp
@@ -101,11 +102,12 @@ install-ui: build-ui
 	$(call LINK_LAB_EXTENSION,services)
 	$(call LINK_LAB_EXTENSION,ui-components)
 	$(call LINK_LAB_EXTENSION,metadata-common)
+	$(call LINK_LAB_EXTENSION,script-editor)
 	$(call INSTALL_LAB_EXTENSION,theme)
 	$(call INSTALL_LAB_EXTENSION,code-snippet)
 	$(call INSTALL_LAB_EXTENSION,metadata)
 	$(call INSTALL_LAB_EXTENSION,pipeline-editor)
-	$(call INSTALL_LAB_EXTENSION,script-editor)
+	$(call INSTALL_LAB_EXTENSION,python-editor)
 
 install: install-server install-ui ## Build and install
 	jupyter lab build
@@ -144,7 +146,7 @@ dist-ui: build-ui
 	$(call PACKAGE_LAB_EXTENSION,code-snippet)
 	$(call PACKAGE_LAB_EXTENSION,metadata)
 	$(call PACKAGE_LAB_EXTENSION,pipeline-editor)
-	$(call PACKAGE_LAB_EXTENSION,script-editor)
+	$(call PACKAGE_LAB_EXTENSION,python-editor)
 
 release: dist-ui build-server ## Build wheel file for release
