+++ packages/script-editor/src/index.ts
@@ -16,311 +16,7 @@
 
 import '../style/index.css';
 
-import { pyIcon, rIcon } from '@elyra/ui-components';
-
-import {
-  JupyterFrontEnd,
-  JupyterFrontEndPlugin,
-  ILayoutRestorer
-} from '@jupyterlab/application';
-import { WidgetTracker, ICommandPalette } from '@jupyterlab/apputils';
-import { CodeEditor, IEditorServices } from '@jupyterlab/codeeditor';
-import { IDocumentWidget } from '@jupyterlab/docregistry';
-import { IFileBrowserFactory } from '@jupyterlab/filebrowser';
-import { FileEditor, IEditorTracker } from '@jupyterlab/fileeditor';
-import { ILauncher } from '@jupyterlab/launcher';
-import { IMainMenu } from '@jupyterlab/mainmenu';
-import { ISettingRegistry } from '@jupyterlab/settingregistry';
-
-import { JSONObject } from '@lumino/coreutils';
-
-import { ScriptEditorFactory, ScriptEditor } from './ScriptEditor';
-
-const SCRIPT_EDITOR_NAMESPACE = 'elyra-script-editor-extension';
-
-const PYTHON_FACTORY = 'Python Editor';
-const PYTHON = 'python';
-const PYTHON_EDITOR_NAMESPACE = 'elyra-python-script-editor-extension';
-
-const R_FACTORY = 'R Editor';
-const R = 'r';
-const R_EDITOR_NAMESPACE = 'elyra-r-script-editor-extension';
-
-const commandIDs = {
-  createNewPythonFile: 'script-editor:create-new-python-file',
-  createNewRFile: 'script-editor:create-new-r-file',
-  openDocManager: 'docmanager:open',
-  newDocManager: 'docmanager:new-untitled'
-};
-
-/**
- * Initialization data for the script editor extension.
- */
-const extension: JupyterFrontEndPlugin<void> = {
-  id: SCRIPT_EDITOR_NAMESPACE,
-  autoStart: true,
-  requires: [
-    IEditorServices,
-    IEditorTracker,
-    ICommandPalette,
-    ISettingRegistry,
-    IFileBrowserFactory
-  ],
-  optional: [ILayoutRestorer, IMainMenu, ILauncher],
-  activate: (
-    app: JupyterFrontEnd,
-    editorServices: IEditorServices,
-    editorTracker: IEditorTracker,
-    palette: ICommandPalette,
-    settingRegistry: ISettingRegistry,
-    browserFactory: IFileBrowserFactory,
-    restorer: ILayoutRestorer | null,
-    menu: IMainMenu | null,
-    launcher: ILauncher | null
-  ) => {
-    console.log('Elyra - script-editor extension is activated!');
-
-    app.docRegistry.addFileType({
-      name: PYTHON,
-      displayName: 'Python File',
-      extensions: ['.py'],
-      pattern: '.*\\.py$',
-      mimeTypes: ['text/x-python'],
-      icon: pyIcon
-    });
-
-    app.docRegistry.addFileType({
-      name: R,
-      displayName: 'R File',
-      extensions: ['.r'],
-      pattern: '.*\\.r$',
-      mimeTypes: ['text/x-rsrc'],
-      icon: rIcon
-    });
-
-    const pythonFactory = new ScriptEditorFactory({
-      editorServices,
-      factoryOptions: {
-        name: PYTHON_FACTORY,
-        fileTypes: [PYTHON],
-        defaultFor: [PYTHON]
-      }
-    });
-
-    const rFactory = new ScriptEditorFactory({
-      editorServices,
-      factoryOptions: {
-        name: R_FACTORY,
-        fileTypes: [R],
-        defaultFor: [R]
-      }
-    });
-
-    const { restored } = app;
-
-    /**
-     * Track Editor widget on page refresh
-     */
-    const pythonEditorTracker = new WidgetTracker<ScriptEditor>({
-      namespace: PYTHON_EDITOR_NAMESPACE
-    });
-    const rEditorTracker = new WidgetTracker<ScriptEditor>({
-      namespace: R_EDITOR_NAMESPACE
-    });
-
-    let config: CodeEditor.IConfig = { ...CodeEditor.defaultConfig };
-
-    if (restorer) {
-      // Handle state restoration
-      void restorer.restore(pythonEditorTracker, {
-        command: commandIDs.openDocManager,
-        args: widget => ({
-          path: widget.context.path,
-          pythonFactory: PYTHON_FACTORY
-        }),
-        name: widget => widget.context.path
-      });
-
-      void restorer.restore(rEditorTracker, {
-        command: commandIDs.openDocManager,
-        args: widget => ({
-          path: widget.context.path,
-          rFactory: R_FACTORY
-        }),
-        name: widget => widget.context.path
-      });
-    }
-
-    /**
-     * Update the setting values. Adapted from fileeditor-extension.
-     */
-    const updateSettings = (settings: ISettingRegistry.ISettings): void => {
-      config = {
-        ...CodeEditor.defaultConfig,
-        ...(settings.get('editorConfig').composite as JSONObject)
-      };
-
-      // Trigger a refresh of the rendered commands
-      app.commands.notifyCommandChanged();
-    };
-
-    /**
-     * Update the settings of the current tracker instances. Adapted from fileeditor-extension.
-     */
-    const updateTracker = (): void => {
-      pythonEditorTracker.forEach(widget => {
-        updateWidget(widget);
-      });
-
-      rEditorTracker.forEach(widget => {
-        updateWidget(widget);
-      });
-    };
-
-    /**
-     * Update the settings of a widget. Adapted from fileeditor-extension.
-     */
-    const updateWidget = (widget: any): void => {
-      if (!editorTracker.has(widget)) {
-        (editorTracker as WidgetTracker<IDocumentWidget<FileEditor>>).add(
-          widget
-        );
-      }
-
-      const editor = widget.content.editor;
-      Object.keys(config).forEach((keyStr: string) => {
-        const key = keyStr as keyof CodeEditor.IConfig;
-        editor.setOption(key, config[key]);
-      });
-    };
-
-    // Notify the widget tracker if restore data needs to update
-    const handleWidgetRestore = (widget: any, tracker: any): void => {
-      void tracker.add(widget);
-
-      widget.context.pathChanged.connect(() => {
-        void tracker.save(widget);
-      });
-      updateWidget(widget);
-    };
-
-    // Fetch the initial state of the settings. Adapted from fileeditor-extension.
-    Promise.all([
-      settingRegistry.load('@jupyterlab/fileeditor-extension:plugin'),
-      restored
-    ])
-      .then(([settings]) => {
-        updateSettings(settings);
-        updateTracker();
-        settings.changed.connect(() => {
-          updateSettings(settings);
-          updateTracker();
-        });
-      })
-      .catch((reason: Error) => {
-        console.error(reason.message);
-        updateTracker();
-      });
-
-    app.docRegistry.addWidgetFactory(pythonFactory);
-    app.docRegistry.addWidgetFactory(rFactory);
-
-    pythonFactory.widgetCreated.connect((sender, widget) => {
-      handleWidgetRestore(widget, pythonEditorTracker);
-    });
-    rFactory.widgetCreated.connect((sender, widget) => {
-      handleWidgetRestore(widget, rEditorTracker);
-    });
-
-    // Handle the settings of new widgets. Adapted from fileeditor-extension.
-    pythonEditorTracker.widgetAdded.connect((sender, widget) => {
-      updateWidget(widget);
-    });
-    rEditorTracker.widgetAdded.connect((sender, widget) => {
-      updateWidget(widget);
-    });
-
-    /**
-     * Create new script file from launcher and file menu
-     */
-
-    if (launcher) {
-      launcher.add({
-        command: commandIDs.createNewPythonFile,
-        category: 'Elyra',
-        rank: 2
-      });
-
-      launcher.add({
-        command: commandIDs.createNewRFile,
-        category: 'Elyra',
-        rank: 3
-      });
-    }
-
-    if (menu) {
-      menu.fileMenu.newMenu.addGroup(
-        [{ command: commandIDs.createNewPythonFile }],
-        30
-      );
-
-      menu.fileMenu.newMenu.addGroup(
-        [{ command: commandIDs.createNewRFile }],
-        31
-      );
-    }
-
-    // Function to create a new untitled script file, given the current working directory
-    const createNew = async (
-      cwd: string,
-      ext: string,
-      factory: any
-    ): Promise<any> => {
-      const model = await app.commands.execute(commandIDs.newDocManager, {
-        path: cwd,
-        type: 'file',
-        ext: ext
-      });
-      return app.commands.execute(commandIDs.openDocManager, {
-        path: model.path,
-        factory: factory
-      });
-    };
-
-    // Add a command to create new Python file
-    app.commands.addCommand(commandIDs.createNewPythonFile, {
-      label: args => (args['isPalette'] ? 'New Python File' : 'Python File'),
-      caption: 'Create a new Python file',
-      icon: args => (args['isPalette'] ? undefined : pyIcon),
-      execute: args => {
-        const cwd = args['cwd'] || browserFactory.defaultBrowser.model.path;
-        return createNew(cwd as string, '.py', PYTHON_FACTORY);
-      }
-    });
-
-    // Add a command to create new R file
-    app.commands.addCommand(commandIDs.createNewRFile, {
-      label: args => (args['isPalette'] ? 'New R File' : 'R File'),
-      caption: 'Create a new R file',
-      icon: args => (args['isPalette'] ? undefined : rIcon),
-      execute: args => {
-        const cwd = args['cwd'] || browserFactory.defaultBrowser.model.path;
-        return createNew(cwd as string, '.r', R_FACTORY);
-      }
-    });
-
-    palette.addItem({
-      command: commandIDs.createNewPythonFile,
-      args: { isPalette: true },
-      category: 'Elyra'
-    });
-
-    palette.addItem({
-      command: commandIDs.createNewRFile,
-      args: { isPalette: true },
-      category: 'Elyra'
-    });
-  }
-};
-
-export default extension;
+export * from './KernelDropDown';
+export * from './ScriptEditor';
+export * from './ScriptEditorController';
+export * from './ScriptRunner';
